Add the basic build toolchain
Makefile downloads the sources of the Iosevka font and the Nerd Fonts
patcher script, builds the custom version of Iosevka and then patches it.

Currently the `patch-fonts.mk` script is buggy and considers the
resulting patched build to be permanently dirty due to a mismatch of
the real output filenames from the expected ones.